requirements.txt
docker-py==1.2.1
redis==2.10.3
